requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
stripe==10.5.0
requests==2.32.3

